// src/api/generated/models/adminUserResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * ASMS API
 * Assignment Management System - REST API
 * OpenAPI spec version: v1
 */
import type { AdminUserResponseLanguage } from './adminUserResponseLanguage'
import type { AdminUserResponseStatus } from './adminUserResponseStatus'
import type { AdminUserResponseSystemRole } from './adminUserResponseSystemRole'

export interface AdminUserResponse {
  id: string
  email: string
  fullName: string
  userCode: string
  systemRole: AdminUserResponseSystemRole
  status: AdminUserResponseStatus
  language: AdminUserResponseLanguage
  createdAt: string
  activatedAt?: string
  lastLoginAt?: string
}