// src/api/generated/models/adminUserResponseStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * ASMS API
 * Assignment Management System - REST API
 * OpenAPI spec version: v1
 */

export type AdminUserResponseStatus =
  (typeof AdminUserResponseStatus)[keyof typeof AdminUserResponseStatus]

export const AdminUserResponseStatus = {
  PENDING_ACTIVATION: 'PENDING_ACTIVATION',
  ACTIVE: 'ACTIVE',
  LOCKED: 'LOCKED',
} as const